Move the effect routing support into the core proper
